Refactored the directory structure
