+++ frontend/src/components/Admin/AddDeviceBody.js
@@ -4,7 +4,7 @@ import {Button, Form} from 'react-bootstrap';
 import { makeStyles } from "@material-ui/core/styles";
 import { useTheme } from "@material-ui/core/styles";
 import componentStyles from "assets/theme/views/admin/admin";
-import "../../views/admin/modals/addDevice.css";
+import "../../assets/css/addDevice.css";
 // import AdminModal from 'admin-modal.js'
