+++ frontend/src/components/Admin/AddParkBody.js
@@ -3,7 +3,7 @@ import React,{useState} from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import { useTheme } from "@material-ui/core/styles";
 import componentStyles from "assets/theme/views/admin/admin";
-import "../../views/admin/modals/addDevice.css";
+import "../../assets/css/addDevice.css";
