+++ frontend/src/components/Admin/ParkTable.js
@@ -1,4 +1,4 @@
-import React,{useState, useEffect} from "react";
+import React, {useEffect, useState} from "react";
 
 import { makeStyles } from "@material-ui/core/styles";
 import { useTheme } from "@material-ui/core/styles";
@@ -11,43 +11,38 @@ import Table from "@material-ui/core/Table";
 import TableBody from "@material-ui/core/TableBody";
 import TableCell from "@material-ui/core/TableCell";
 import TableContainer from "@material-ui/core/TableContainer";
-import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
 import Typography from "@material-ui/core/Typography";
 import componentStyles from "assets/theme/views/admin/admin";
 import Button from "@material-ui/core/Button";
-import Modal from "../../views/admin/modals/Modal";
-import AddUserBody from "./AddUserBody";
-// import disableScroll from 'disable-scroll';
+import Modal from "../Modals/Modal";
+import disableScroll from "disable-scroll";
+import AddParkBody from "./AddParkBody";
 
-// import AdminModal from 'admin-modal'
 
 
 const useStyles = makeStyles(componentStyles);
 
-const UserTable = () => {
+const ParkTable = () => {
     const classes = useStyles();
     const theme = useTheme();
     const [show, setShow] = useState(false);
-
-    // useEffect(() => {
-    //     if (show == true) disableScroll.on()
-    //     if (show == false) disableScroll.off()
-    // }, [show])
+    useEffect(() => {
+        if (show == true) disableScroll.on()
+        if (show == false) disableScroll.off()
+    }, [show])
 
     return (
         <>
             <Container
                 maxWidth={false}
                 component={Box}
-                marginTop="-6rem"
                 classes={{ root: classes.containerRoot }}
             >
-                <Modal title="Add User" onClose={() => setShow(false)} show={show}>
-                    <AddUserBody/>
+                <Modal title="Add Park" onClose={() => setShow(false)} show={show}>
+                    <AddParkBody/>
                 </Modal>
-
-                <Grid container component={Box} marginTop="3rem">
+                <Grid container component={Box} >
                     <Grid
                         item
                         xs={12}
@@ -72,7 +67,7 @@ const UserTable = () => {
                                                 variant="h2"
                                                 marginBottom="0!important"
                                             >
-                                                Users
+                                                Parks
                                             </Box>
                                         </Grid>
                                         <Grid item xs="auto">
@@ -86,73 +81,14 @@ const UserTable = () => {
                                     </Grid>
                                 }
                                 classes={{ root: classes.cardHeaderRoot }}
-
-                            >
-                                </CardHeader>
+                            ></CardHeader>
                             <TableContainer>
                                 <Box
                                     component={Table}
                                     alignItems="center"
                                     marginBottom="0!important"
                                 >
-                                    <TableHead>
-                                        <TableRow style={{background: 'rgb(243 243 243)'}}>
-                                            <TableCell
-                                                classes={{
-                                                    root:
-                                                        classes.tableCellRoot +
-                                                        " " +
-                                                        classes.tableCellRootHead,
-                                                }}
-                                            >
-                                                User Name
-                                            </TableCell>
-                                            <TableCell
-                                                classes={{
-                                                    root:
-                                                        classes.tableCellRoot +
-                                                        " " +
-                                                        classes.tableCellRootHead,
-                                                }}
-                                            >
-                                                Role
-                                            </TableCell>
-                                            <TableCell
-                                                classes={{
-                                                    root:
-                                                        classes.tableCellRoot +
-                                                        " " +
-                                                        classes.tableCellRootHead,
-                                                }}
-                                            >
-                                                Park
-                                            </TableCell>
-                                            <TableCell
-                                                classes={{
-                                                    root:
-                                                        classes.tableCellRoot +
-                                                        " " +
-                                                        classes.tableCellRootHead,
-                                                }}
-                                            >
-
-                                            </TableCell>
-                                            <TableCell
-                                                classes={{
-                                                    root:
-                                                        classes.tableCellRoot +
-                                                        " " +
-                                                        classes.tableCellRootHead,
-                                                }}
-                                            >
-
-                                            </TableCell>
-                                        </TableRow>
-                                    </TableHead>
                                     <TableBody>
-
-
-                                        {/*default values:*/}
                                         <TableRow>
                                             <TableCell
                                                 classes={{
@@ -162,16 +98,8 @@ const UserTable = () => {
                                                         classes.tableCellRootBodyHead,
                                                 }}
                                                 scope="row"
-                                                style={{verticalAlign:'middle',width:'25%'}}
+                                                style={{verticalAlign:'middle', width:'80%'}}
                                             >
-                                                Jane Doe
-                                            </TableCell>
-                                            <TableCell classes={{ root: classes.tableCellRoot }}
-                                                       style={{verticalAlign:'middle',width:'20%'}}>
-                                                Ranger
-                                            </TableCell>
-                                            <TableCell classes={{ root: classes.tableCellRoot }}
-                                                       style={{verticalAlign:'middle',width:'34%'}}>
                                                 Riet Vlei
                                             </TableCell>
                                             <TableCell classes={{ root: classes.tableCellRoot }}
@@ -215,7 +143,7 @@ const UserTable = () => {
                         component={Box}
                         marginBottom="3rem!important"
                         classes={{ root: classes.gridItemRoot }}
-                        style={{ display: "flex", justifyContent: "flex-end", marginTop: "5px"}}>
+                        style={{ display: "flex", justifyContent: "flex-end", marginTop: "5px" }}>
                         <Button
                             variant="contained"
                             color="primary"
@@ -223,7 +151,7 @@ const UserTable = () => {
                             style={{width:'200px'}}
                             onClick={() => setShow(true)}
                         >
-                            Add User
+                            Add Park
                         </Button>
                     </Grid>
                 </Grid>
@@ -233,4 +161,4 @@ const UserTable = () => {
     );
 };
 
-export default UserTable;
+export default ParkTable;
