+++ frontend/src/components/Dashboard/DeviceDetails.js
