+++ frontend/src/components/Dashboard/DeviceTable.js
