+++ frontend/src/components/Dashboard/InspectionTable.js
