+++ frontend/src/components/Headers/Header.js
@@ -14,7 +14,7 @@ import InsertChartOutlined from "@material-ui/icons/InsertChartOutlined";
 import PieChart from "@material-ui/icons/PieChart";
 
 // core components
-import CardStats from "components/Cards/CardStats.js";
+import CardStats from "components/Dashboard/CardStats.js";
 
 import componentStyles from "assets/theme/components/header.js";
 import axios from "axios";
