+++ frontend/src/components/Modals/Modal.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import ReactDOM from "react-dom";
 import { CSSTransition } from "react-transition-group";
-import "./modal.css";
+import "../../assets/css/modal.css";
 import Button from "@material-ui/core/Button";
 
 const Modal = props => {
