+++ frontend/src/views/admin/Admin.js
@@ -1,13 +1,13 @@
 import React from "react";
 import AdminHeader from "../../components/Headers/AdminHeader";
-import UserTable from "../../components/Custom/UserTable";
+import UserTable from "../../components/Admin/UserTable";
 import Grid from "@material-ui/core/Grid";
 import Box from "@material-ui/core/Box";
 import componentStyles from "assets/theme/views/dashboard/dashboard.js";
 import { makeStyles } from "@material-ui/core/styles";
 import Container from "@material-ui/core/Container";
-import DeviceTable from "../../components/Custom/DeviceTable";
-import ParkTable from "../../components/Custom/ParkTable";
+import DeviceTable from "../../components/Admin/DeviceTable";
+import ParkTable from "../../components/Admin/ParkTable";
 
 
 const useStyles = makeStyles(componentStyles);
